Create kidsWithCandies.java
problem is here:

https://leetcode.com/problems/kids-with-the-greatest-number-of-candies/?envType=study-plan-v2&envId=leetcode-75


my solution thought process:

first i go through the entire list 1 time to find the highest number of candies. this will take O(n) time to do since we go through n items 1 time.

then again we go through the entire list and compare the numbers 1 at a time and the comparitor we use is to add the number we have/are on with the extraCandies. if this value is more than OR EQUAL TO the max we had from 1st run though the list, then we will add true to our arraylist we will be returning at the end. else we add a false boolean instead.

this will take O(n) time again.

so we are at 2 times O(n) time complexity for all of it but 2 is a constant so that can be ignored. hence whole algo time complexity is O(n).