Create UniqueOccurrences.java
problem link:

https://leetcode.com/problems/unique-number-of-occurrences/?envType=study-plan-v2&envId=leetcode-75

hashmap/hashset solution.

simply map the occurences. key:value -> number:occurences

then add each VALUE to a set. if any value occurs more than 1 time, then it wont get added to set. then return a comparison of the size of hashmap to the size of the hashset. if not the same size, then that means a value was repeated and not added to hashset. else size will be equal if all occurence counts/value were unique.